commit before moving job->run logic out onto the scripts responsible for running the jobs so we can accurately call shutdown functions without creating massive memory leaks (since workers would create new shutdown functions for each job ran, but never clear any of the job objects because they would hold a reference to in the shutdown function, which never gets called)

+++ src/Dren/FileLockableDataStore.php
@@ -22,11 +22,11 @@ class FileLockableDataStore implements LockableDataStore
         // Update: I'm told that file locks are released whenever script execution completes, but I'm not entirely
         // sure that's correct...and what does that mean even? Like, when the script exits or when gc runs? Leaving
         // this here because it makes me feel better.
-        register_shutdown_function(function() {
-
-            $this->closeLock();
-
-        });
+//        register_shutdown_function(function() {
+//
+//            $this->closeLock();
+//
+//        });
 
     }
 
@@ -207,7 +207,8 @@ class FileLockableDataStore implements LockableDataStore
 
     public function deleteUnsafe(): void
     {
-        unlink($this->fileFullPath);
+        if(file_exists($this->fileFullPath))
+            unlink($this->fileFullPath);
     }
 
     public function deleteUnsafeById(string $id) : void
